third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/scatterND.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] scatterND 1D float16 tensors (Insert individual elements in a tensor by index)
  assert_less_than_equal: assert_array_approx_equals_ulp: test scatterND float16 actual 2.00390625 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 1026
[FAIL] scatterND 3D float16 tensors (Insert entire slices of a higher rank tensor)
  assert_less_than_equal: assert_array_approx_equals_ulp: test scatterND float16 actual 2052 should be close enough to expected 5 by ULP distance: expected a number less than or equal to 0 but got 8962
Harness: the test ran to completion.
